・64bit版GCC4.7.2 on FedoraCore18 でエラーとなる処理を修正中（一旦コミット）。

+++ include/gasha/hash_table.cpp.h
@@ -732,7 +732,7 @@ namespace hash_table
 		if (index == INVALID_INDEX)//検索失敗なら削除失敗
 			return false;
 		_eraseByIndex(index);
-		if (m_usingCount == m_deletedCount || m_deletedCount == AUTO_REHASH_SIZE)//自動リハッシュ
+		if (m_usingCount == m_deletedCount || m_deletedCount == static_cast<int>(AUTO_REHASH_SIZE))//自動リハッシュ
 			_rehash();
 		return true;
 	}
